Layout and logic

+++ android/TypeOneTwo/src/com/appspot/typeonetwo/models/Insulin.java
@@ -6,10 +6,10 @@ public class Insulin {
     public static final int RAPIDACTING = 0;
     public static final int LONGACTING = 1;
     int type;
-    float dose;
+    double dose;
     Date date;
 
-    public Insulin(int type, float dose, Date date) {
+    public Insulin(int type, double dose, Date date) {
         this.type = type;
         this.dose = dose;
         this.date = date;
